use multi-prefix verify instead of file splitting
This eliminates all preprocessor-based test suppression (SUPPRESS_OUT_OF_BOUND
and UNINIT_WITHOUT_OUTOFBOUND) without introducing new files.

Three verify prefix groups are used:
- 'oob': expectations only checked when OutOfBounds is enabled (sinks)
- 'no-oob': expectations only checked when OutOfBounds is disabled
- 'uninit': expectations only checked when UninitializedRead is enabled
